- initial import of IBM xl compiler support to allow check of distribution - m.c.
Subversion: revision:407

+++ src/adapters/compiler/SILC_compiler_pgi.c
@@ -56,7 +56,7 @@ struct s1
     char*             rout; /* routine name */
 };
 
-unsigned int count = 0;
+uint64_t count = 0;
 
 /**
  * static variable to control initialize status of GNU
